requirements.txt
selenium
request
pillow